requirements.txt
python-telegram-bot==20.4
fastapi==0.100.1
uvicorn==0.23.2
requests==2.28.2
deta==1.2.0
